[MOVE/#1] manifest 등 import 경로 수정
