+++ app/src/main/AndroidManifest.xml
@@ -19,26 +19,25 @@
         android:usesCleartextTraffic="true"
         tools:targetApi="31">
         <activity
-            android:name=".SignUpActivity"
-            android:exported="false">
-        </activity>
+            android:name=".presentation.signUp.SignUpActivity"
+            android:exported="false"></activity>
         <activity
-            android:name=".MainHomeActivity"
+            android:name=".presentation.mainhome.MainHomeActivity"
             android:exported="false">
 
         </activity>
         <activity
-            android:name=".LoginActivity"
-            android:exported="true" >
-        <!--Could not identify launch Activity: Default Activity not found 에러 fix목적으로 추가-->
-        <!--추가한 코드 start-->
-        <intent-filter>
-            <action android:name="android.intent.action.MAIN" />
+            android:name=".presentation.login.LoginActivity"
+            android:exported="true">
+            <!--Could not identify launch Activity: Default Activity not found 에러 fix목적으로 추가-->
+            <!--추가한 코드 start-->
+            <intent-filter>
+                <action android:name="android.intent.action.MAIN" />
 
-            <category android:name="android.intent.category.LAUNCHER" />
-        </intent-filter>
-        <!--end-->
-        <!---->
+                <category android:name="android.intent.category.LAUNCHER" />
+            </intent-filter>
+            <!--end-->
+            <!---->
         </activity>
 
     </application>
